Added ProgramText class.
I wasn't sure about the class split at first, and I'm still not really.
However, if the program keeps going in this direction (and it looks like
it will for a while), then this is a fairly significant step, because it
clears the way for a lot of other stuff. The ProgramText class is now
the thing that's interacting with the backend and displaying things to
the user. The Node class has become a very, very basic canvas layout
engine. One good way to continue the program, and the way I think I'm
going to pursue now, is to expand Node's capabilities to include laying
out whole trees of windows, probably through helper classes.