Anchored layout system.
This commit begins the anchored layout system, which is a new way of
arranging Nodes on a Canvas. In this system, Nodes are arranged by a
Layout object (currently either a row or a column layout) that
determines their correct positions relative to other Nodes. Each layout
is anchored by a Node, whose placement determines the placement of every
other node in the layout, but which is not actually placed by the
layout. In this way, every node on the canvas is in a tree in which
higher nodes determine the placement of lower nodes.

The system is full of flaws, and honestly, I think the fundamental idea
is bad, but it's good enough to display sombrero data right now. Later
on maybe someone will write a better layout system.