Read traces better.
This mades trace reading a bit nicer, including (finally!) printing
clarifications next to defs and uses so you can see which is which.

The big issue in reading right now is that the definition of a variable
appears both as the result of the use and as a child of the use. The
fundamental issue is that an ExpConstUse node in a trace file doesn't
actually store a result - it only knows the ExpConstDef it references.
If we were assuming eager evaluation, it would be easy to just return
the final result of the def as the result of the use, but that's
inelegant - the trace should record this information itself.

I believe Hat's current strategy is to just return the final result,
except that the final result of a Haskell program isn't necessarily a
value - it might be an unentered application.

The strategy I'm moving towards right now is the function subNodes,
based on the traverseReduction function in HatExplore.hs in the Hat
source - it basically returns a pair with the subexpressions that a
given node actually creates and the constdefs that it accesses. This
gives us the flexibility to show const defs as only results, but it
doesn't really solve the issue of what to count as the result of an
ExpConstUse (unless we're fine letting the result be the def).