Arrows and hiding.
Layouts now draw arrows on the screen to show which nodes are connected
to which, a vast improvement over the previous strategy of doing
nothing. This also starts work on letting the user hide nodes if they're
not interesting, but it still contains bugs, and I still don't know what
the right thing to do is when the user hides a node that has visible
children.