Improve build system.
This patch changes the build system to handle SWIG modules better.
First, the names of the extensions are changed in setup.py so that
distutils builds correctly-named shared libraries (there's a note in
setup.py with more information). Then, it adds a hack to copy everything
in a build/lib* directory to a toplevel (in the working tree) directory
called 'lib'. This puts well-named shared libraries in a consistent
location, which means we can modify sombrero to add lib in its
PYTHONPATH variable and not have to manually copy libraries any more.

This also adds a note in the TODO file about something I noticed in the
Python documentation, and it excludes SWIG-generated Python files from
git (but you have to manually add them to the list).