Added .gitignore
Today I learned about .gitignore files, and realized that it would be
good for this project to have one. This means that everyone who clones
this repository will get my list of files for git not to track along
with everything else. This should be quite useful for people who want to
develop Sombrero themselves.