Minor gui work.
Do a bit more for the gui, and also change trace-interpret.py to work
with Python 2.5 (at the cost of breaking it for 2.6). Maybe someday I'll
learn how to make things conditional on the version of Python I'm using.
Or if not, I'll just wait until 2.6 is out of beta and then drop 2.5
support.