Fix trace reading bug.
The trace reader was crashing because of an issue that ultimately turned
out to be the misuse of peekExpArg in artutils.c - we were calling it on
the expression 2, which is not legal, and then there was some sort of
strange error, probably in the SWIG wrapper, which hid the fact that it
had reached an assert(False) statement. A change to program.py fixes
that, and a comment there suggests a better fix for the future.