GUI bug fixes.
There are now no known GUI bugs! The layout stuff is working (as far as
I can tell) and nodes are moving themselves around when other nodes are
added to the graph. I'm sure more bugs will appear in time, and
certainly it'll need more features, but for now, this is a good enough
foundation that it's time to put some more backend work on top of it.